fix(gesture): prepend origin in reconstruct; widen total_us; spread test
Whole-branch review findings:
- C1 (Critical): reconstruct misaligned pts[0] position vs fraction, collapsing
  knots and zeroing leading dt_q for large-first-delta (flick) gestures. Now
  prepends the implicit origin (0,0)@t=0 as a real path point so the first leg
  spreads correctly; first-leg duration assumed one nominal interval.
- I1 (Important): added flick spread/dt_q regression test.
- I2 (Important): total_us widened to uint32_t (was saturating at ~256 ms).

+++ src/gesture.h
@@ -25,7 +25,7 @@ typedef struct {
     gst_knot_t knots[GST_KNOTS_MAX];
     uint16_t   n;             /* knots used (<= GST_KNOTS_MAX)            */
     float      raw_len;       /* original path length (counts)           */
-    uint16_t   total_us;      /* original duration (x nominal)           */
+    uint32_t   total_us;      /* original duration, .8 fixed of nominal (no 16-bit cap) */
     uint8_t    submv;         /* submovements detected                   */
     uint8_t    flags;         /* curvature sign, click-adjacent, etc.    */
 } gst_shape_t;
